reverted status refactoring
